Boilerplate on base bottom sheet
