+++ app/src/main/java/com/p2p/presentation/home/join/JoinGamesBottomSheetFragment.kt
@@ -1,16 +1,9 @@
 package com.p2p.presentation.home.join
 
-import android.app.Dialog
-import android.os.Bundle
 import android.view.LayoutInflater
-import android.view.View
 import android.view.ViewGroup
-import android.widget.FrameLayout
 import androidx.fragment.app.viewModels
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.google.android.material.bottomsheet.BottomSheetBehavior
-import com.google.android.material.bottomsheet.BottomSheetBehavior.BottomSheetCallback
-import com.google.android.material.bottomsheet.BottomSheetDialog
 import com.p2p.databinding.FragmentJoinGamesBinding
 import com.p2p.framework.bluetooth.BluetoothDeviceFinderReceiver
 import com.p2p.presentation.base.BaseBottomSheetDialogFragment
@@ -25,6 +18,7 @@ class JoinGamesBottomSheetFragment :
     override val viewModel: JoinGamesViewModel by viewModels { JoinGamesViewModelFactory(receiver) }
     override val inflater: (LayoutInflater, ViewGroup?, Boolean) -> FragmentJoinGamesBinding =
         FragmentJoinGamesBinding::inflate
+    override val isCollapsable = false
 
     override fun initUI() {
         setupRecycler()
@@ -35,29 +29,6 @@ class JoinGamesBottomSheetFragment :
     override fun setupObservers() = with(viewModel) {
         devices.observe(viewLifecycleOwner) { adapter.devices = it.toList() }
         connectButtonEnabled.observe(viewLifecycleOwner) { binding.connectButton.isEnabled = it }
-    }
-
-    override fun onCreateDialog(savedInstanceState: Bundle?): Dialog {
-        return super.onCreateDialog(savedInstanceState).apply {
-            setOnShowListener {
-                val dialog = it as BottomSheetDialog
-                val bottomSheet = dialog.findViewById<FrameLayout>(com.google.android.material.R.id.design_bottom_sheet)
-                    ?: return@setOnShowListener
-                BottomSheetBehavior.from(bottomSheet).apply {
-                    state = BottomSheetBehavior.STATE_EXPANDED
-                    addBottomSheetCallback(object : BottomSheetCallback() {
-                        override fun onStateChanged(bottomSheet: View, newState: Int) {
-                            if (newState == BottomSheetBehavior.STATE_COLLAPSED) {
-                                state = BottomSheetBehavior.STATE_EXPANDED
-                            }
-                        }
-
-                        override fun onSlide(bottomSheet: View, slideOffset: Float) {
-                        }
-                    })
-                }
-            }
-        }
     }
 
     override fun onEvent(event: JoinGamesEvent) = when (event) {
