+++ app/src/main/res/values/strings.xml
@@ -16,7 +16,7 @@
 
     <!-- Join games -->
     <string name="join_games_title">Seleccioná el celular de quien creó el juego</string>
-    <string name="join_game_troubleshoot">En caso de no encontrar el celular, deben conectarse manualmente. Hace click acá y seguí las instrucciones.</string>
+    <string name="join_game_troubleshoot">En caso de no encontrar el celular deben conectarse manualmente. Hacé click acá y sigan las instrucciones en ambos celulares.</string>
 
     <!-- Games names -->
     <string name="games_tutti_frutti">Tutti Frutti</string>
